Remove force type casting on getSkin()

+++ Forge1.12.2/src/main/java/com/Fishmod/mod_LavaCow/entities/EntityFoglet.java
@@ -322,7 +322,7 @@ public class EntityFoglet extends EntityMob implements IAggressive{
     
     public int getSkin()
     {
-        return ((Integer)this.dataManager.get(SKIN_TYPE)).intValue();
+        return this.dataManager.get(SKIN_TYPE).intValue();
     }
 
     public void setSkin(int skinType)
