+++ Forge1.12.2/src/main/java/com/Fishmod/mod_LavaCow/entities/flying/EntityPtera.java
@@ -186,7 +186,7 @@ public class EntityPtera extends EntityFlyingMob {
    
    public int getSkin()
    {
-       return ((Integer)this.dataManager.get(SKIN_TYPE)).intValue();
+       return this.dataManager.get(SKIN_TYPE).intValue();
    }
 
    public void setSkin(int skinType)
